重构：Pet，Pet继承了Item，却又作为Item属性使用。直接使用Pet，PetId通过YitIdHelper生成

+++ src/Application.Core.EF/Entities/Items/PetEntity.cs
@@ -2,7 +2,7 @@ namespace Application.EF.Entities;
 
 public class PetEntity
 {
-    public PetEntity(int petid, string? name, int level, int closeness, int fullness, bool summoned, int flag)
+    public PetEntity(long petid, string? name, int level, int closeness, int fullness, bool summoned, int flag)
     {
         Petid = petid;
         Name = name;
@@ -14,7 +14,7 @@ public class PetEntity
     }
 
     protected PetEntity() { }
-    public int Petid { get; set; }
+    public long Petid { get; set; }
 
     public string? Name { get; set; }
